chore: fix wrapping errors for all pkgs
