+++ pkg/builder/docker/errors.go
@@ -1,30 +1,10 @@
 package docker
 
 import (
-	"fmt"
+	"github.com/celestiaorg/knuu/pkg/errors"
 )
 
-type Error struct {
-	Code    string
-	Message string
-	Err     error
-}
-
-func (e *Error) Error() string {
-	if e.Err == e {
-		return e.Message
-	}
-
-	if e.Err != nil {
-		return fmt.Sprintf("%s: %v", e.Message, e.Err)
-	}
-	return e.Message
-}
-
-func (e *Error) Wrap(err error) error {
-	e.Err = err
-	return e
-}
+type Error = errors.Error
 
 var (
 	ErrFailedToListBuildxBuilders = &Error{Code: "FailedToListBuildxBuilders", Message: "failed to list buildx builders"}
