+++ pkg/builder/kaniko/errors.go
@@ -1,30 +1,10 @@
 package kaniko
 
 import (
-	"fmt"
+	"github.com/celestiaorg/knuu/pkg/errors"
 )
 
-type Error struct {
-	Code    string
-	Message string
-	Err     error
-}
-
-func (e *Error) Error() string {
-	if e.Err == e {
-		return e.Message
-	}
-
-	if e.Err != nil {
-		return fmt.Sprintf("%s: %v", e.Message, e.Err)
-	}
-	return e.Message
-}
-
-func (e *Error) Wrap(err error) error {
-	e.Err = err
-	return e
-}
+type Error = errors.Error
 
 var (
 	ErrBuildFailed                      = &Error{Code: "BuildFailed", Message: "build failed"}
