+++ pkg/traefik/errors.go
@@ -1,34 +1,10 @@
 package traefik
 
 import (
-	"errors"
-	"fmt"
+	"github.com/celestiaorg/knuu/pkg/errors"
 )
 
-type Error struct {
-	Code    string
-	Message string
-	Err     error
-	Params  []interface{}
-}
-
-func (e *Error) Error() string {
-	msg := fmt.Sprintf(e.Message, e.Params...)
-	if e.Err != nil {
-		return fmt.Sprintf("%s: %v", msg, e.Err)
-	}
-	return msg
-}
-
-func (e *Error) Wrap(err error) error {
-	e.Err = errors.Join(e.Err, err)
-	return e
-}
-
-func (e *Error) WithParams(params ...interface{}) *Error {
-	e.Params = params
-	return e
-}
+type Error = errors.Error
 
 var (
 	ErrTraefikDeploymentCreationFailed   = &Error{Code: "TraefikDeploymentCreationFailed", Message: "error creating Traefik deployment"}
